refactor: Carousel.tsx 를 관심사에 따라 컴포넌트 분리함
